@@ -1,5 +1,5 @@
 name: libdbm
-version: 0.1.5
+version: 0.1.6
 description: A tiny and fast (1000s read/write per second) database with
   a simple API (`Map`) for Dart using disk-based hashing. No dependencies,
   batteries included.
@@ -8,13 +8,13 @@ repository: https://github.com/libdbm/libdbm
 issue_tracker: https://github.com/libdbm/libdbm/issues
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
 
 dev_dependencies:
-  faker: ^1.3.0
-  pedantic: ^1.9.0
-  test: ^1.14.4
-  collection: ^1.14.13
+  faker: ^2.0.0-rc.1
+  pedantic: ^1.11.0
+  test: ^1.16.7
+  collection: ^1.15.0
 
