@@ -8,11 +8,10 @@ import 'package:libdbm/src/util.dart';
 void main() {
   File file = File('dummy.bin');
   setUp(() async {
-    if (file.existsSync()) file.deleteSync(recursive: true);
-    file.createSync(recursive: true);
+    if (file.existsSync()) try { file.deleteSync(recursive: true); } finally {}
   });
   tearDown(() async {
-    if (file.existsSync()) file.deleteSync(recursive: true);
+    if (file.existsSync()) try { file.deleteSync(recursive: true); } finally {}
   });
   test('Test alignment', () {
     expect(align(123, 128), equals(128));
